requirements.txt
fastapi
uvicorn[standard]
neo4j
pydantic
python-dotenv
pytest
httpx
jupyter
pandas